requirements.txt
# Core Streamlit and web framework
streamlit>=1.28.0
requests>=2.31.0

# Web searching and scraping
ddgs>=0.3.0
newspaper3k>=0.2.8
beautifulsoup4>=4.12.0
lxml>=4.9.0

# AI and machine learning
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
torch>=2.0.0
transformers>=4.30.0

# Google Gemini API
google-generativeai>=0.3.0

# OCR and PDF processing
easyocr
pdf2image
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
tqdm>=4.65.0

# Optional: For enhanced functionality
langchain>=0.0.200